# api/docs/swagger.yaml
definitions:
  models.Book:
    properties:
      author_id:
        type: string
      category_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Category:
    properties:
      name:
        type: string
      parent_id:
        type: string
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
  models.ListBooks:
    properties:
      Books:
        items:
          $ref: '#/definitions/models.Book'
        type: array
    type: object
  models.ListCategories:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
    type: object
  models.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
  models.UpdateBook:
    properties:
      author_id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/books:
    get:
      consumes:
      - application/json
      description: This API for getting list of books
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      - description: Author
        in: query
        name: author
        type: string
      - description: Category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListBooks'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: ListBooks
      tags:
      - book
  /v1/books/:
    post:
      consumes:
      - application/json
      description: This API for creating a new book
      parameters:
      - description: bookCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateBook
      tags:
      - book
  /v1/books/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting book
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteBook
      tags:
      - book
    get:
      consumes:
      - application/json
      description: This API for getting book detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetBook
      tags:
      - book
    put:
      consumes:
      - application/json
      description: This API for updating book
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: BookUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBook'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateBook
      tags:
      - book
  /v1/categories:
    get:
      consumes:
      - application/json
      description: This API for getting list of categories
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListCategories'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: ListCategories
      tags:
      - category
  /v1/categories/:
    post:
      consumes:
      - application/json
      description: This API for creating a new category
      parameters:
      - description: categoryCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateCategory
      tags:
      - category
  /v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting category
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteCategoryById
      tags:
      - category
    get:
      consumes:
      - application/json
      description: This API for getting category detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetCategoryById
      tags:
      - category
    put:
      consumes:
      - application/json
      description: This API for updating category
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: categoryUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateCategory
      tags:
      - category
swagger: "2.0"